refactor delete button

+++ videobookmarks/static/index.js
@@ -3,13 +3,15 @@ const deleteButtons = document.getElementsByClassName('button big delete');
 var link;
 for (button of deleteButtons){
     button.addEventListener('click', function (e) {
-        link = this.value;
-        fetch(link, {
-            method: 'DELETE',
-            headers: {
-              'Content-Type': 'application/json',
-            }
-        });
-        location.reload();
+        if (window.confirm('Are you sure you want to delete this lookmark collection?')) {
+            link = this.value;
+            fetch(link, {
+                method: 'DELETE',
+                headers: {
+                  'Content-Type': 'application/json',
+                }
+            });
+            location.reload();
+        };
     });
 };
